Covert returned transaction types
