+++ src/adapter/web/batch.mjs
@@ -4,7 +4,7 @@ import {
   unwrapData,
   updateHelpers,
   writeHelpers
-} from './index.mjs'
+} from './core.mjs'
 
 export const batch = (rootDB, options) => {
   assertEnvironment(options?.as)
